+++ sharding-core/src/main/java/io/shardingsphere/core/rule/DataSourceParameter.java
@@ -30,6 +30,8 @@ import lombok.Setter;
 @Setter
 public final class DataSourceParameter {
     
+    public static final String DATA_SOURCE_POOL_CLASS_NAME = "com.zaxxer.hikari.HikariDataSource";
+    
     private ProxyPoolType proxyDatasourceType = ProxyPoolType.VENDOR;
     
     private String url;