+++ sharding-orchestration/sharding-orchestration-core/src/test/java/io/shardingsphere/orchestration/internal/state/AllStateTests.java
@@ -17,7 +17,9 @@
 
 package io.shardingsphere.orchestration.internal.state;
 
-import io.shardingsphere.orchestration.internal.state.listener.AllListenerTests;
+import io.shardingsphere.orchestration.internal.state.listener.DataSourceStateOrchestrationListenerTest;
+import io.shardingsphere.orchestration.internal.state.listener.InstanceStateOrchestrationListenerTest;
+import io.shardingsphere.orchestration.internal.state.listener.StateOrchestrationListenerManagerTest;
 import io.shardingsphere.orchestration.internal.state.node.StateNodeTest;
 import io.shardingsphere.orchestration.internal.state.service.DataSourceServiceTest;
 import io.shardingsphere.orchestration.internal.state.service.InstanceStateServiceTest;
@@ -30,7 +32,9 @@ import org.junit.runners.Suite.SuiteClasses;
         StateNodeTest.class,
         InstanceStateServiceTest.class,
         DataSourceServiceTest.class,
-        AllListenerTests.class
+        StateOrchestrationListenerManagerTest.class, 
+        DataSourceStateOrchestrationListenerTest.class,
+        InstanceStateOrchestrationListenerTest.class
 })
 public final class AllStateTests {
 }