+++ sharding-proxy/src/main/java/io/shardingsphere/shardingproxy/backend/SchemaBroadcastBackendHandler.java
@@ -22,11 +22,9 @@ import io.shardingsphere.shardingproxy.backend.jdbc.connection.BackendConnection
 import io.shardingsphere.shardingproxy.runtime.GlobalRegistry;
 import io.shardingsphere.shardingproxy.transport.common.packet.DatabasePacket;
 import io.shardingsphere.shardingproxy.transport.mysql.packet.command.CommandResponsePackets;
-import io.shardingsphere.shardingproxy.transport.mysql.packet.generic.ErrPacket;
 import io.shardingsphere.shardingproxy.transport.mysql.packet.generic.OKPacket;
 import lombok.RequiredArgsConstructor;
 
-import java.sql.SQLException;
 import java.util.Collection;
 import java.util.LinkedList;
 import java.util.List;
@@ -45,19 +43,17 @@ public final class SchemaBroadcastBackendHandler implements BackendHandler {
     
     private final String sql;
     
+    private final BackendConnection backendConnection;
+    
     private final DatabaseType databaseType;
     
     @Override
     public CommandResponsePackets execute() {
         List<DatabasePacket> packets = new LinkedList<>();
         for (String schema : GlobalRegistry.getInstance().getSchemaNames()) {
-            try (BackendConnection backendConnection = new BackendConnection()) {
-                BackendHandler backendHandler = BackendHandlerFactory.newTextProtocolInstance(connectionId, sequenceId, sql, backendConnection, databaseType, schema);
-                CommandResponsePackets commandResponsePackets = backendHandler.execute();
-                packets.addAll(commandResponsePackets.getPackets());
-            } catch (final SQLException ex) {
-                return new CommandResponsePackets(new ErrPacket(1, ex));
-            }
+            BackendHandler backendHandler = BackendHandlerFactory.newTextProtocolInstance(connectionId, sequenceId, sql, backendConnection, databaseType, schema);
+            CommandResponsePackets commandResponsePackets = backendHandler.execute();
+            packets.addAll(commandResponsePackets.getPackets());
         }
         return merge(packets);
     }