+++ sharding-proxy/src/test/java/io/shardingsphere/shardingproxy/frontend/common/executor/ExecutorGroupTest.java
@@ -72,7 +72,7 @@ public final class ExecutorGroupTest {
     
     private ShardingProperties getShardingProperties(final TransactionType transactionType) {
         Properties props = new Properties();
-        props.setProperty(ShardingPropertiesConstant.PROXY_TRANSACTION_ENABLED.getKey(), String.valueOf(transactionType == TransactionType.XA));
+        props.setProperty(ShardingPropertiesConstant.PROXY_TRANSACTION_TYPE.getKey(), transactionType.name());
         return new ShardingProperties(props);
     }
 }