+++ sharding-transaction/sharding-transaction-xa/src/test/java/io/shardingsphere/transaction/xa/manager/AtomikosTransactionManagerTest.java
@@ -68,7 +68,7 @@ public final class AtomikosTransactionManagerTest {
     }
     
     @After
-    public void teardown() {
+    public void tearDown() {
         ReflectiveUtil.setProperty(atomikosTransactionManager, "underlyingTransactionManager", underlyingTransactionManager);
         atomikosTransactionManager.destroy();
     }