+++ RELEASE-NOTES.md
@@ -14,12 +14,14 @@
 
 #### Core
 
+1. [ISSUE #1348](https://github.com/sharding-sphere/sharding-sphere/issues/1348) Exception occur when using PreparedStatement to run SELECT LIMIT in multi datasources
 1. [ISSUE #1422](https://github.com/sharding-sphere/sharding-sphere/issues/1422) Exception occur when using defaultDataSourceName and defaultDatabaseStrategy at same time
 
 #### Sharding-Proxy
 
 1. [ISSUE #1403](https://github.com/sharding-sphere/sharding-sphere/issues/1403) master-slave mode exception when using PreparedStatement
 1. [ISSUE #1414](https://github.com/sharding-sphere/sharding-sphere/issues/1414) `SHOW CREATE TABLE tb_name` execution exception when configuring multiple logic schema
+1. [ISSUE #1475](https://github.com/sharding-sphere/sharding-sphere/issues/1475) `SHOW [FUll] TABLES FROM db_name LIKE 'tb_name'` failed to return the logic table
 
 ## 3.0.0